Add conditional rendering for show all/show previous, and moved booked container to containers
